произведено деление на модули
