+++ package.json
@@ -5,7 +5,7 @@
   "description": "Личный проект «Кекстаграм» от HTML Academy",
   "repository": {
     "type": "git",
-    "url": "git@github.com:htmlacademy-javascript/2131405-kekstagram-26.git"
+    "url": "git@github.com:htmlacademy-javascript/2131405-kekstagram-26.git",
   },
   "bugs": {
     "url": "https://github.com/htmlacademy-javascript/2131405-kekstagram-26/issues"
@@ -24,5 +24,5 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  }
+  },
 }
